Conflicts:
	src/mongo/bson/bsondemo/bsondemo.cpp

+++ src/mongo/dbtests/documenttests.cpp
@@ -662,7 +662,7 @@ namespace DocumentTests {
                 intrusive_ptr<const Value> value() { return Value::createDate(0); }
             };
             
-            /** Coerce // to bool. */
+            /** Coerce js literal regex to bool. */
             class RegexToBool : public ToBoolTrue {
                 intrusive_ptr<const Value> value() { return fromBson( fromjson( "{''://}" ) ); }
             };
@@ -805,7 +805,7 @@ namespace DocumentTests {
             class LongToDouble : public ToDoubleBase {
                 intrusive_ptr<const Value> value() {
                     // A long that cannot be exactly represented as a double.
-                    return Value::createDouble( 0x8fffffffffffffffLL );
+                    return Value::createDouble( static_cast<double>( 0x8fffffffffffffffLL ) );
                 }
                 double expected() { return static_cast<double>( 0x8fffffffffffffffLL ); }
             };
